Validando parámetros con class-validator y mapped-types
