+++ package.json
@@ -23,7 +23,10 @@
   "dependencies": {
     "@nestjs/common": "^8.0.0",
     "@nestjs/core": "^8.0.0",
+    "@nestjs/mapped-types": "^1.0.1",
     "@nestjs/platform-express": "^8.0.0",
+    "class-transformer": "^0.5.1",
+    "class-validator": "^0.13.2",
     "eslint-plugin-prettier": "^4.2.1",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
