+++ src/controllers/products/products.controller.ts
@@ -14,7 +14,7 @@ import {
 } from '@nestjs/common';
 // import { Response } from 'express';
 import { ProductsService } from 'src/services/products/products.service';
-import { CreateProductDto, UpdateProductDto } from 'src/dto/products.dtos';
+import { CreateProductDto, UpdateProductDto } from 'src/dtos/products.dtos';
 import { ParseIntPipe } from 'src/common/parse-int.pipe';
 @Controller('products')
 export class ProductsController {
