--- src/dto/products.dtos.ts
@@ -1,14 +1,0 @@
-export class CreateProductDto {
-	readonly name: string;
-	readonly description: string;
-	readonly price: number;
-	readonly stock: number;
-	readonly image: string;
-}
-export class UpdateProductDto {
-	readonly name?: string;
-	readonly description?: string;
-	readonly price?: number;
-	readonly stock?: number;
-	readonly image?: string;
-}
